Don't get fooled by non-testing variables
Only calls to t.Run and t.Parallel should count. Without this, the
linter might see a call to some other Run function and then complain
that it contains no call to Parallel inside.

This works by capturing the name of the testing.T parameter and only
recognizing calls to Run and Parallel on variables with that name. False
positives might occur if the parameter gets hidden.

Fixed #9